+++ src/classes/nft.js
@@ -1,14 +1,8 @@
-const Ajv = require("ajv");
-const { ErrorObject } = require("ajv");
-const ajv = new Ajv(); // options can be passed, e.g. {allErrors: true}
-const config = require('getconfig');
-const fs = require('fs');
-const fs_ = require('fs/promises');
-const path = require('path');
-const JSONschemaDefaults = require('json-schema-defaults');
 
-const { fileExists } = require('../utils/helpers.js');
+// const { fileExists } = require('../utils/helpers.js');
 const { promptSchema } = require('../utils/prompt.js');
+const { fromSchema, loadSchemaFromFile, validate } = require("../utils/schema.js");
+
 const Asset = require('./asset.js');
 const IPFS = require('./ipfs.js');
 
@@ -58,12 +52,10 @@ class NFT {
     
         if (process.env.cli)
             this.schemaJSON = await promptSchema(this.schema);
-        else if (this.schema.includes("ipfs"))
-            this.schemaJSON = await NFT.loadSchemaFromIPFS(this.schema);
         else
-            this.schemaJSON = await NFT.loadSchemaFromFile(this.schema);
+            this.schemaJSON = await loadSchemaFromFile(this.schema);
 
-        this.metadata = NFT.fromSchema(this.schemaJSON);
+        this.metadata = fromSchema(this.schemaJSON);
 
         this._initialized = true;
     }
@@ -77,63 +69,18 @@ class NFT {
     // or interface with contract.methods.supportsInterface or whatever the function is
     // static getStandard() {}
 
-    // TODO
-    // create NFT from metadata {} object possibly with schema name
-    // static fromMetadata(metadata) {return {}}
-
-    // TODO
-    // load directly from metadata at ipfs
-    static fromIPFS() {}
-
-    // TODO
-    // return object from schema
-    static fromSchema(schema) {return JSONschemaDefaults(schema)}
-
     // return schema as an IPFS cid if available
-    static _getSchemaCID(schema) {
-        for (const [key, value] in config.schemasIPFS)
-            if (key == schema) return config.schemasIPFS[key];
-        return null;
-    }
+    // static _getSchemaCID(schema) {
+    //     for (const [key, value] in config.schemasIPFS)
+    //         if (key == schema) return config.schemasIPFS[key];
+    //     return null;
+    // }
 
-    // TODO
-    // return config.schemasIPFS if available
-    static _getSchemasCID() {
-        return config.schemasIPFS || null;
-    }
-
-    // TODO
-    // load schema from ipfs
-    static async loadSchemaFromIPFS(schema) {
-        let cid = NFT._getSchemaCID(schema);
-        // if (!cid) cid = NFT._getSchemasCID();
-        return null;
-
-    }
-
-    // locally from files
-    static loadSchemaFromFile(schema) {
-        console.debug(`loading schema: ${schema}...`);
-        function _parseTemplate(t) {return JSON.parse(fs.readFileSync(`${config.SCHEMA_PATH}/${t}.json`))}
-        // get list of template files from available files in available /schema directories
-        const templates = [];
-        const localPath = path.join(__dirname, "../..", config.SCHEMA_PATH), // path local to this script
-              addonPath = path.join(process.env.PWD, config.SCHEMA_PATH);// path to where the cwd is
-        const files = [];
-        if (fileExists(localPath)) 
-            files.push(...fs.readdirSync(localPath));
-        if (addonPath != localPath && fileExists(addonPath))
-            files.push(...fs.readdirSync(addonPath));
-        let defaultIndex = 0;
-        for (let i=0;i<files.length;i++) {
-            const filename = files[i].replace(".json","");
-            if (filename === schema) return _parseTemplate(filename);
-            templates.indexOf(filename) === -1 ? templates.push(filename) : console.debug(`duplicate template found: ${filename}`)
-            // set simple.json to default template
-            if (filename === "simple") defaultIndex = i;
-        }
-        return _parseTemplate(templates[defaultIndex]);
-    }
+    // // TODO
+    // // return config.schemasIPFS if available
+    // static _getSchemasCID() {
+    //     return config.schemasIPFS || null;
+    // }
 
     ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
     ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -175,7 +122,7 @@ class NFT {
 
     async upload() {
         if (!this._initialized) await this.init();
-        this.validate();
+        validate(this.metadata, this.schema, this.schemaJSON);
         await Asset.uploadAssets(this.metadata);
         await this.uploadMetadata();
     }
@@ -190,21 +137,6 @@ class NFT {
         const { metadataCID, metadataURI } = await IPFS.add(file);
         this.metadataCID = metadataCID;
         this.metadataURI = metadataURI;
-    }
-
-    // validate according to its schema
-    validate() {
-        console.debug(`validating NFT schema: ${this.schema}...`);
-        // replace empty values with null for flagging validation
-        for (const [key, value] of Object.entries(this.metadata))
-            if (value === "") this.metadata[key] = null;
-        const validate = ajv.compile(this.schemaJSON);
-        const valid = validate(this.metadata);
-        if (!valid) {
-            console.error(validate.errors);
-            throw "Error: unable to validate data";
-        }
-        console.debug("valid JSON!");
     }
 
 }