+++ src/utils/crypto.js
@@ -1,0 +1,219 @@
+
+// import * as fs from 'fs';
+// import * as path from 'path';
+// import * as crypto from 'crypto';
+
+const fs = require("fs");
+const path = require("path");
+const crypto = require("crypto");
+
+generateKeys()
+
+export async function encryptFile(file) {
+  try {
+    const name = path.basename(file);
+    const buff = fs.readFileSync(file);
+    const key = crypto.randomBytes(16).toString('hex'); // 16 bytes -> 32 chars
+    const iv = crypto.randomBytes(8).toString('hex');   // 8 bytes -> 16 chars
+    const ekey = encryptRSA(key); // 32 chars -> 684 chars
+    const ebuff = encryptAES(buff, key, iv);
+
+    const content = Buffer.concat([ // headers: encrypted key and IV (len: 700=684+16)
+      Buffer.from(ekey, 'utf8'),   // char length: 684
+      Buffer.from(iv, 'utf8'),     // char length: 16
+      Buffer.from(ebuff, 'utf8')
+    ])
+
+    console.log('ENCRYPTION --------')
+    console.log('key:', key, 'iv:', iv, 'ekey:', ekey.length)
+    console.log('contents:', buff.length, 'encrypted:', ebuff.length)
+    console.log(' ')
+
+    return { content, key, name };
+  } catch (err) {
+    console.log(err)
+    throw err;
+  }
+}
+
+export async function encryptFolder(folderPath) {
+  // console.log("folderPath:", folderPath);
+  const files = fs.readdirSync(folderPath);
+  const encryptedFiles = [];
+  for (const file of files) {
+    // console.log("file:", file)
+    encryptedFiles.push(await encryptFile(path.join(folderPath, file)));
+  }
+  return encryptedFiles;
+}
+
+export async function decryptFile(file_data) {
+  try {    
+    let edata = []
+    for await (const chunk of file_data)
+      edata.push(chunk)
+    edata = Buffer.concat(edata)
+    const key = decryptRSA(edata.slice(0, 684).toString('utf8'))
+    const iv = edata.slice(684, 700).toString('utf8')
+    const econtent = edata.slice(700).toString('utf8')
+    const ebuf = Buffer.from(econtent, 'hex')
+    const content = decryptAES(ebuf, key, iv)
+    console.log(' ')
+    console.log('DECRYPTION --------')
+    console.log('key:', key, 'iv:', iv)
+    console.log('contents:', content.length, 'encrypted:', econtent.length)
+    console.log('downloaded:', edata.length)
+    return content
+  } catch (err) {
+    console.log(err)
+    throw err;
+  }
+}
+
+////////////////////////////////
+//////////// IPFS //////////////
+////////////////////////////////
+
+async function uploadFileEncrypted(file, ipfs) {
+  try {
+    const ipfspath = '/encrypted/data/' + path.basename(file) // ipfspath
+    const buff = fs.readFileSync(file);
+    const key = crypto.randomBytes(16).toString('hex'); // 16 bytes -> 32 chars
+    const iv = crypto.randomBytes(8).toString('hex');   // 8 bytes -> 16 chars
+    const ekey = encryptRSA(key); // 32 chars -> 684 chars
+    const ebuff = encryptAES(buff, key, iv);
+
+    const content = Buffer.concat([ // headers: encrypted key and IV (len: 700=684+16)
+      Buffer.from(ekey, 'utf8'),   // char length: 684
+      Buffer.from(iv, 'utf8'),     // char length: 16
+      Buffer.from(ebuff, 'utf8')
+    ])
+    
+    // await ipfs.files.write(
+    //   ipfspath,
+    //   content,
+    //   {create: true, parents: true}
+    // );
+
+    const added = await ipfs.add(
+      {
+        content: content
+      }
+    )
+
+    console.log('ENCRYPTION --------')
+    console.log('key:', key, 'iv:', iv, 'ekey:', ekey.length)
+    console.log('contents:', buff.length, 'encrypted:', ebuff.length)
+    console.log(' ')
+
+    const url = `https://freewilly.infura-ipfs.io/ipfs/${added.path}`
+    console.log("Upload successful!");
+    // console.log(added);
+    console.log(url);
+    return added.path;
+  } catch (err) {
+    console.log(err)
+    throw err;
+  }
+}
+
+async function downloadFileEncrypted(ipfspath, ipfs) {
+  try {
+    // let file_data = await ipfs.files.read(ipfspath)
+    let file_data = await ipfs.get(ipfspath);
+    
+    let edata = []
+    for await (const chunk of file_data)
+      edata.push(chunk)
+    edata = Buffer.concat(edata)
+
+    const key = decryptRSA(edata.slice(0, 684).toString('utf8'))
+    const iv = edata.slice(684, 700).toString('utf8')
+    const econtent = edata.slice(700).toString('utf8')
+    const ebuf = Buffer.from(econtent, 'hex')
+    const content = decryptAES(ebuf, key, iv)
+
+    console.log(' ')
+    console.log('DECRYPTION --------')
+    console.log('key:', key, 'iv:', iv)
+    console.log('contents:', content.length, 'encrypted:', econtent.length)
+    console.log('downloaded:', edata.length)
+    
+    return content
+    
+  } catch (err) {
+    console.log(err)
+    throw err;
+  }
+}
+
+
+////////////////////////////////
+
+async function toArray(asyncIterator) { 
+  const arr=[]; 
+  for await(const i of asyncIterator) {
+    arr.push(i); 
+  }
+  return arr;
+}
+
+////////////////////////////////
+
+function encryptAES(buffer, secretKey, iv) {
+  const cipher = crypto.createCipheriv('aes-256-ctr', secretKey, iv);
+  const data = cipher.update(buffer);
+  const encrypted = Buffer.concat([data, cipher.final()]);
+  return encrypted.toString('hex')
+}
+
+function decryptAES(buffer, secretKey, iv) {
+  const decipher = crypto.createDecipheriv('aes-256-ctr', secretKey, iv);
+  const data = decipher.update(buffer)
+  const decrpyted = Buffer.concat([data, decipher.final()]);
+  return decrpyted;
+}
+
+function generateKeys() {
+  if (fs.existsSync('private.pem') && fs.existsSync('public.pem'))
+    return;
+  
+  const { privateKey, publicKey } = crypto.generateKeyPairSync('rsa', {
+    modulusLength: 4096,
+    publicKeyEncoding: {
+      type: 'pkcs1',
+      format: 'pem',
+    },
+    privateKeyEncoding: {
+      type: 'pkcs1',
+      format: 'pem',
+      cipher: 'aes-256-cbc',
+      passphrase: '',
+    },
+  })
+
+  fs.writeFileSync('private.pem', privateKey)
+  fs.writeFileSync('public.pem', publicKey)
+}
+
+function encryptRSA(toEncrypt, pubkeyPath='public.pem') {
+  const absolutePath = path.resolve(pubkeyPath)
+  const publicKey = fs.readFileSync(absolutePath, 'utf8')
+  const buffer = Buffer.from(toEncrypt, 'utf8')
+  const encrypted = crypto.publicEncrypt(publicKey, buffer)
+  return encrypted.toString('base64')
+}
+
+function decryptRSA(toDecrypt, privkeyPath='private.pem') {
+  const absolutePath = path.resolve(privkeyPath)
+  const privateKey = fs.readFileSync(absolutePath, 'utf8')
+  const buffer = Buffer.from(toDecrypt, 'base64')
+  const decrypted = crypto.privateDecrypt(
+  {
+    key: privateKey.toString(),
+    passphrase: '',
+  },
+  buffer,
+  )
+  return decrypted.toString('utf8')
+}